+++ scripting-definition/src/main/kotlin/ScriptDefinition.kt
@@ -17,7 +17,9 @@
 package net.edwardday.serverscript.scriptdefinition
 
 import kotlinx.coroutines.runBlocking
+import net.edwardday.serverscript.scriptdefinition.annotation.Import
 import net.edwardday.serverscript.scriptdefinition.script.ServerScript
+import java.io.File
 import kotlin.script.experimental.annotations.KotlinScript
 import kotlin.script.experimental.api.RefineScriptCompilationConfigurationHandler
 import kotlin.script.experimental.api.ResultWithDiagnostics
@@ -25,11 +27,13 @@ import kotlin.script.experimental.api.ScriptCollectedData
 import kotlin.script.experimental.api.ScriptCompilationConfiguration
 import kotlin.script.experimental.api.ScriptConfigurationRefinementContext
 import kotlin.script.experimental.api.ScriptDiagnostic
+import kotlin.script.experimental.api.ScriptSourceAnnotation
 import kotlin.script.experimental.api.asDiagnostics
 import kotlin.script.experimental.api.asSuccess
 import kotlin.script.experimental.api.collectedAnnotations
 import kotlin.script.experimental.api.defaultImports
 import kotlin.script.experimental.api.implicitReceivers
+import kotlin.script.experimental.api.importScripts
 import kotlin.script.experimental.api.onSuccess
 import kotlin.script.experimental.api.refineConfiguration
 import kotlin.script.experimental.dependencies.CompoundDependenciesResolver
@@ -38,6 +42,8 @@ import kotlin.script.experimental.dependencies.FileSystemDependenciesResolver
 import kotlin.script.experimental.dependencies.Repository
 import kotlin.script.experimental.dependencies.maven.MavenDependenciesResolver
 import kotlin.script.experimental.dependencies.resolveFromScriptSourceAnnotations
+import kotlin.script.experimental.host.FileBasedScriptSource
+import kotlin.script.experimental.host.toScriptSource
 import kotlin.script.experimental.jvm.dependenciesFromClassContext
 import kotlin.script.experimental.jvm.jvm
 import kotlin.script.experimental.jvm.updateClasspath
@@ -53,7 +59,7 @@ abstract class ServerScriptDefinition
 
 class ServerScriptConfiguration : ScriptCompilationConfiguration(
     body = {
-        defaultImports(DependsOn::class, Repository::class)
+        defaultImports(DependsOn::class, Repository::class, Import::class)
         defaultImports("net.edwardday.serverscript.scriptdefinition.script.*")
         jvm {
             dependenciesFromClassContext(
@@ -66,13 +72,73 @@ class ServerScriptConfiguration : ScriptCompilationConfiguration(
         refineConfiguration {
             // the callback called when any of the listed file-level annotations are encountered in the compiled script
             // the processing is defined by the `handler`, that may return refined configuration depending on the annotations
-            onAnnotations(DependsOn::class, Repository::class, handler = ServerScriptConfigurator())
+            onAnnotations(DependsOn::class, Repository::class, handler = ServerScriptClasspathConfigurator())
+            onAnnotations(Import::class, handler = ServerScriptImportsConfigurator())
         }
 
     }
 )
 
-class ServerScriptConfigurator : RefineScriptCompilationConfigurationHandler {
+class ServerScriptImportsConfigurator : RefineScriptCompilationConfigurationHandler {
+    override fun invoke(context: ScriptConfigurationRefinementContext): ResultWithDiagnostics<ScriptCompilationConfiguration> {
+        val annotations = context.collectedData?.get(ScriptCollectedData.collectedAnnotations)
+            ?.takeIf(List<ScriptSourceAnnotation<*>>::isNotEmpty)
+        val importAnnotations = annotations?.filter { it.annotation is Import }?.takeIf(List<*>::isNotEmpty)
+            ?: return context.compilationConfiguration.asSuccess()
+        val workingDir = (context.script as? FileBasedScriptSource)?.file?.absoluteFile?.parentFile
+        val diagnostics = arrayListOf<ScriptDiagnostic>()
+        val scripts = importAnnotations.mapNotNull { (annotation, location) ->
+            val path = (annotation as Import).path.let {
+                if (it.endsWith(".$SERVER_SCRIPT_FILE_EXTENSION")) it else "$it.$SERVER_SCRIPT_FILE_EXTENSION"
+            }
+            val absolutePathFile = File(path).takeIf(File::isAbsolute)
+            when {
+                absolutePathFile != null -> absolutePathFile.takeIf(File::isFile).also {
+                    if (it == null) {
+                        diagnostics += ScriptDiagnostic(
+                            code = ScriptDiagnostic.unspecifiedError,
+                            message = "cannot find import script file at $path",
+                            severity = ScriptDiagnostic.Severity.WARNING,
+                            locationWithId = location,
+                        )
+                    }
+                }
+
+                workingDir != null -> File(workingDir, path).takeIf(File::isFile).also {
+                    if (it == null) {
+                        diagnostics += ScriptDiagnostic(
+                            code = ScriptDiagnostic.unspecifiedError,
+                            message = "cannot find import script file at ${workingDir.path}/$path",
+                            severity = ScriptDiagnostic.Severity.WARNING,
+                            locationWithId = location,
+                        )
+                    }
+                }
+
+                else -> {
+                    diagnostics += ScriptDiagnostic(
+                        code = ScriptDiagnostic.unspecifiedError,
+                        message = "cannot find import script file with relative path ($path) from a script with unknown location",
+                        severity = ScriptDiagnostic.Severity.ERROR,
+                        locationWithId = location,
+                    )
+                    null
+                }
+            }?.toScriptSource()
+        }
+
+        return if (diagnostics.isNotEmpty()) {
+            ResultWithDiagnostics.Failure(diagnostics)
+        } else {
+            ScriptCompilationConfiguration(context.compilationConfiguration) {
+                importScripts(scripts)
+            }.asSuccess()
+        }
+    }
+
+}
+
+class ServerScriptClasspathConfigurator : RefineScriptCompilationConfigurationHandler {
     private val resolver = CompoundDependenciesResolver(FileSystemDependenciesResolver(), MavenDependenciesResolver())
 
     override operator fun invoke(context: ScriptConfigurationRefinementContext): ResultWithDiagnostics<ScriptCompilationConfiguration> =