+++ scripting-definition/src/main/kotlin/annotation/Import.kt
@@ -1,0 +1,6 @@
+package net.edwardday.serverscript.scriptdefinition.annotation
+
+@Target(AnnotationTarget.FILE)
+@Repeatable
+@Retention(AnnotationRetention.SOURCE)
+annotation class Import(val path: String)