+++ scripting-host/src/test/resources/imports/invocation_counter.server.kts
@@ -1,0 +1,4 @@
+
+val sharedInvocation: Int = cache.updateOrSet("INVOCATION", { 1 }, 1::plus)
+
+writeOutput("shared invocation counter: $sharedInvocation\n")